requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
bleach==5.0.1
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.1
django-allauth==0.51.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
drf-api-logger==1.1.11
drf-tracking==1.5.0
drf-yasg==1.21.3
ghp-import==2.1.0
idna==3.4
importlib-metadata==5.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.3.7
MarkupSafe==2.1.1
mergedeep==1.3.4
mkdocs==1.4.2
oauthlib==3.2.1
packaging==21.3
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python-mimeparse==1.6.0
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
pyyaml-env-tag==0.1
requests==2.28.1
requests-oauthlib==1.3.1
RestAuthCommon==0.7.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sentry-sdk==1.9.8
six==1.16.0
sqlparse==0.4.2
types-cryptography==3.3.23
uritemplate==4.1.1
urllib3==1.26.12
watchdog==2.1.9
webencodings==0.5.1
zipp==3.10.0
